Restructure the deploy-contract API route
