+++ client/package.json
@@ -56,6 +56,7 @@
     "@vercel/kv": "^1.0.1",
     "axios": "^1.7.9",
     "bcryptjs": "^3.0.2",
+    "chalk": "^5.6.0",
     "class-variance-authority": "^0.7.1",
     "clsx": "^2.1.1",
     "csstype": "^3.1.3",
