+++ client/types/deploy.ts
@@ -1,0 +1,6 @@
+export interface CompilationResult {
+  success: boolean;
+  contracts: string[];
+  error?: string;
+  errorLog?: string;
+}
